Melhorias percebidas durante implementação da classe em um projeto real.

+++ src/Elements/ICMSIPI/C001.php
@@ -14,7 +14,7 @@ class C001 extends Element implements ElementInterface
 
     protected $parameters = [
         'IND_MOV' => [
-            'type'     => 'integer',
+            'type'     => 'numeric',
             'regex'    => '^[0-1]{1}$',
             'required' => true,
             'info'     => 'Indicador de movimento: '
