+++ src/Elements/ICMSIPI/Z0001.php
@@ -23,7 +23,7 @@ class Z0001 extends Element implements ElementInterface
     
     protected $parameters = [
         'ind_mov' => [
-            'type'     => 'integer',
+            'type'     => 'numeric',
             'regex'    => '^[0-1]{1}$',
             'required' => true,
             'info'     => 'Indicador de movimento: 0- Bloco com dados informados; 1- Bloco sem dados informados.',
